Rewrite dataset_definition for debugging prior to loading
Rewrite the dataset definition to a temporary file before passing it
to the unsafe loader (which doesn't have permission to write). This
means we can truncate it at a stop() call, but it can still can use
the same module loading code as a normal dataset definition, and
imports from other modules will still work.