fix: Include all patients in PatientColumn derived from filtered EventColumn
This commit follows on from e0aa72c.

Previously, if you filtered an event series such that not every patient
had a record in the new series, and then picked one record per patient
from the new event series, then the resulting patient series would not
include those patients:

    >>> clinical_events.where(...).sort_by(clinical_events.date).first_for_patient().date
    patient_id        | value
    ------------------+------------------
    3                 | 2023-04-17
    4                 | 2023-03-13
    5                 | 2023-02-04
    6                 | 2022-06-19
    7                 | 2022-08-30
    8                 | 2024-02-08
    10                | 2023-06-21

But now:

    >>> clinical_events.where(...).sort_by(clinical_events.date).first_for_patient().date
    patient_id        | value
    ------------------+------------------
    1                 | None
    2                 | None
    3                 | 2023-04-17
    4                 | 2023-03-13
    5                 | 2023-02-04
    6                 | 2022-06-19
    7                 | 2022-08-30
    8                 | 2024-02-08
    9                 | None
    10                | 2023-06-21

This reduces the scope for confusion when debugging and when doing the
quiz, since two PatientSeries could have the same values for all patients,
but have different default values.  This means the two series aren't
equal, despite having (before this commit) the same repr.

For instance, these two series have the same value for every patient
when both s and t are not null:

    x = (s < t) | (t.is_null() & s.is_not_null())
    y = s.is_not_null() & (t.is_null() | (s < t))

But x has default None and y has default False.

So if we have the following data in a table called e:

      |  i1 |  i2 |  b1
    --+-----+-----+-----
    1 |  2  |  1  |  T
    2 |  2  |  1  |  F

And with:

    s = e.where(e.b1).sort_by(e.i1).first_for_patient().i1
    t = e.where(e.b2).sort_by(e.i1).first_for_patient().i2

And:

    x = (s > t) | (s.is_null() & t.is_not_null())
    y = ((s > t) | s.is_null()) & t.is_not_null()

Before this change, both x and y would have rendered the same:

    patient_id        | value
    ------------------+------------------
    1                 | True

But now they render differently:

    x:
    patient_id        | value
    ------------------+------------------
    1                 | True
    2                 | None

    y:
    patient_id        | value
    ------------------+------------------
    1                 | True
    2                 | False

They have different defaults because when two patient series are
combined with an operator, the default of the new patient series is the
result of applying the operator to the two defaults.  From
in_memory_database.apply_function:

        return PatientColumn(
            {p: fn(*[col[p] for col in columns]) for p in patients},
            default=fn(*[col.default for col in columns]),
        )

So (and at this stage I'm just trying to convince myself) take:

    a = s > t              default: null
    b = s.is_null()        default: true
    c = t.is_not_null()    default: false

Then:

    x = a | (b & c)        default: null
    y = (a | b) & c        default: false