fix: Make dummy data timeout configurable
We're seeing cases where the generator is making progress, but slow
enough that the default timeout doesn't generate enough patients. Being
able to bump up the timeout would unblock users in this particular case,
although it's obviously not a great solution.

As part of this we use the default values from the config class to
define the keyword argument defaults, and we remove the default values
from the docstring (as they're already shown in the docs in the function
signature). These changes reduce the possibility for these values to go
out of sync.