fix: Render DateDifference elements with other related series
DateDifference nodes are not themselves a Series. However, we
render them as the repr of their .days method, which IS a Series.
In order to determine whether a DateDifference is related to other
debugged elements, we need to compare it as days.

We also move the decision about what to render to the sandbox
query engine's evaluate method.