fix: protect standard out when loading ehrql dataset
Users may reasonably use print() statsments in their dataset definition.
Currently, because we use stdout to output the serialization from the
isolated subprocess, this will break that process by mixing their output
with the json.

So, this change forces any use of sys.stdout within the dataset
definition to use stderr instead, to avoid that happening.