Defer some imports to make sandbox work without any dependencies.
This defers some sqlalchemy imports and some graphviz imports in
ehrql/main.py, in order not to block running the sandbox if they are
missing.

The primary motiviation was to support running the sandbox in pyodide,
but is also just a useful property for ehrql core not to require any
dependencies.

Have not added tests that error handling works as expected, as we may
not move foward with this.