fix: Update tests to add boilerplate for ehrql dataset docs examples
Adds the minimal dataset boilerplace (if it's missing) to create a dataset
and define a population so that an ehrql example can be validated. This
means that the tutorial examples, which mostly use the show() function,
don't need to include the dataset configuration and can still be tested
to ensure they're valid ehrql.