Remove invalid test case
This passed a non-boolean series to a function only defined to accept
booleans. In production code, series have already been validated as
boolean before this function is invoked.

This happens to pass at the moment because the current code doesn't
throw an error here and just silently returns False. But this will
change shortly and we'll get an error.