Fix up links to point to Markdown files
This stops the OpenSAFELY documentation link checker from failing,
because it's able to locate the Markdown file. (But it cannot find a
relative link from documentation source files.)

This can also stop MkDocs from complaining if the appropriate warnings about
relative links are configured.

It also makes the paths make more sense in the context of the
documentation source: the relative paths now point to the appropriate
source file path.