Run `generate-docs`
The metavar names here aren't really user-facing in the sense that they
don't change the API at all, although they do appear in the docs. If we
could be bothered we could keep these the same while changing the
underlying variable name, but I don't think there's anything to be
gained by that.