Add dedicated console output handler
This means we now support writing multiple tables of output to the
console, as we'll need to once `generate-dataset` produces more than one
output table.

At present we just re-use the CSV renderer for this which means that
there's no user-facing change. I think in future we should consider
using some kind of pretty-printer to make the output more user-friendly
and readable.