fix: Fix `CastToInt` operation on Trino engine
The bug was introduced as part of #2359 and then (amazingly!) identified
by the Hypothesis CI run on a totally unrelated PR.

The regression test here is probably unnecessary in that in seems
vanishingly unlikely that we'd introduce this specific bug again. But I
just feel weird fixing something without adding a test which verifies
it.

I think the wider lesson here is that the column types SQLAlchemy gives
us aren't necessarily that reliable: there seem to be quite a few
operations which don't preserve type information.