fix: Reject patient-domained sorts in the query model
These are always a no-op as, necessarily, the value is the same for
every row belonging to a given patient. It wasn't immediately obvious
that it was worth making these an error however. But the issue addressed
in #2371 demonstrates that there's an easily made mistake here that we
need to guard against.