fix: Ensure that None is always a possible dummy data value
If None is an option for possible values, it is usually
given equal weighting to any other value (this is something we might
want to change more generally). In most cases, even if we're reducing
the subset of possible values we select from, we do want None to be one
of the options.