Weight date of death dummy data towards None values
For date of death, if None is a possible value, we usually want it
to also be the most likely value to appear. We treat date of death as
a special case in `choose_random_value` and weight the choices so that
None should be chosen on average 90% of the time.