feat: Prefetch measure variables that do not vary accross intervals
Measures fetch results for each defined interval; if variables
don't vary over the intervals (determined by the presence of the
interval parameter in the variable definition), we fetch them once in
advance of the intervals, and use the results to create an
InlinePatientTable.
We then replace the unvarying variables in the dataset definition
with simple SelectColumns from the inline table.