Move measures dummy data unit tests
This just moves the unit tests for the legacy dummy measures data
and locates them with the other legacy dummy data tests. This is
a) consistent with the layout of the nextgen dummy data tests and
b) makes it clearer what dummy data these tests are testing.

Also removes the DummyMeasuresDataGenerator import from
measures/__init__.py, adds it to dummy_data/__init__.py and cleans
up some imports so that it's clearer which version of the dummy
data you're using.