fix: Use global temporary table for MSSQL results
This allows us to use a separate connection for downloading results
which can safely reconnect and resume on error without losing the
temporary table. This gives us the robustness we used to get when
writing results to a persistent table but without the locking and
performance problems which that gave us (and with the benefit of
automatic cleanup should a job get interrupted).

See PR description for further details.