Fix and simplify gentest randomisation toggle
The comment correctly described the behaviour we wanted, but the code
logic was backwards. This is easy enough to do when you've got an env
var called "RANDOMIZE" toggling a config value called "derandomize" so
we switch the name of the env var to match.

This makes it clear that there's exactly one circumstance where we don't
want randomisation and that's in the CI tests.

Related Slack thread:
https://bennettoxford.slack.com/archives/C069YDR4NCA/p1739440049238879