Use GROUPING SETS and GROUPING ID for measures
To combine a set of compatible measures (i.e. those sharing a
denominator) into a single query.

For mssql we use GROUPING SETS and GROUPING ID. For other sql engines
we reproduce the same behaviour by generating the grouping ID and
doing a UNION ALL on the individual measure queries.