Use SQLAlchemy annotations to track query types
I've experimented with different ways of passing this information around
and come to the conclusion that this is probably the cleanest.

This information is only relevant to the MSSQL engine at the moment as
that's the only one which has to use different approaches to fetching
the results depending on their type. So making the query_type a thing
that every other bit of code needs to be aware of and pass around ends
up adding complexity.

The MSSQL code will blow up loudly if given a query which is not
annotated with its type, so I'm not worried about failing to set the
annotation and not realising.