fix: Refactor setup/cleanup query handling
The big change here is to do away with the three lists representation of
our queries in the form:

    setup_queries, results_queries, cleanup_queries

and instead have a flat list of queries paired with a boolean indicating
whether that query is expected to return any results e.g

    False, sqlalchemy.text("CREATE TABLE ...")
    False, sqlalchemy.text("INSERT INTO ...")
    True,  sqlalchemy.text("SELECT * FROM ...")
    False, sqlalchemy.text("DROP TABLE ...")

This generally simplifies the code and also allows us to arbitrarily
interleave query types so that we're not restricted to doing cleanup
only after all the results queries have run. This will allow us to do
much longer sequences of queries (e.g all the queries required for a set
of measures) without fear of running out of temporary table space.

This does mean that our logs no longer explicitly distinguish between
setup and cleanup queries, but I don't think this is any great loss:
given that we log the query itself it's pretty obvious what sort it is.