Refactor GroupedSum to hold a collection of measures sharing a denominator
Now that we build a single query for a collection of measures that
share a denominator, we don't need to return the denominator alongside
each numerator in the results. The grouping sets (or union_all for
non-mssql) can return one denominator column that's grouped by the
right things for that numerator.