feat: refactor measures calculation to use results from grouped_sum
Replace the fetchers with a GroupedSum object that holds the information
about the numerators, denominator and group-by columns for each measure
(in a collection of measures that share the same denominator).

By using grouping sets, we now return just one table in the list of
results tables, and the rows in the results are now already
summed and grouped.