fix: Fix quiz error checking for series defaults
This attempts to address the confusing behaviour identified in:
 * https://github.com/opensafely/ehrql-tutorial/issues/24

The issue was that although the proposed series gave the right answer
for all patients contained within the series, it would not give the
right answer when combined with series containing other patients.

This is because the final clause:

    | ((has_moderate_or_severe.is_null()) & (has_mild.is_null())))

means the value default to True for missing patients whereas it should
default to False.

Preventing this kind of confusion was the aim of a previous change to
the in-memory engine, but unfortunately it turns out not to be easy to
do this without breaking things elsewhere. See:
 * https://github.com/opensafely-core/ehrql/pull/2256

The new error message is not as absolutely clear as we might like, but
it's an improvement on the status quo.