Make group_bys on GroupedSum a dict of group bys to numerators
This will allow group bys that are shared across measures in a
measure group (measurs sharing a denominator) to be handled
together.

Note that query engines that use grouping sets don't need to
know which numerators go with which group bys, but we will need
this for sqlite, which needs to do separate group by queries and
union all them.