Refactor measures calculator to handle shared group bys
There's not much that needs to change to handle shared group bys.
We need to store a record of group bys and the numerators for the
measures that use them, and for each measure, we store a record
of the position that its numerator will be in in the retrieved
results row.

For convenience, adds a MeasureFetcher dataclass to hold the
measure and its related fetching info.