Add a test for shared group bys
Add a failing test for measures with shared group bys. Currently
this will fail on the assertion that len(list(results)) ==
len(set(results)), because it will return duplicate rows in the
results.  Note that existing tests wouldn't have caught this
because they only assert with set(), so the len assert is also
added to the big main happy-path get_measure_results test (this
passes because there are no measures in the test data that share
both denominiator and group bys).