fix: Don't attempt to fetch measure results in batches
Batched fetching was introduced to work around the unreliability of the
database connection as it allows us to retry part way through the
download without starting again from the beginning. The problem is
particularly acute for patient and event-level results which number in
the millions and where there is an obvious column (`patient_id`) to page
on. However this is less of an issue for aggregated results
(specifically measures) as there are fewer rows to deal with, and these
also lack a natural paging column.

Attempting to use batched fetching on measures results is now causing
issues (see linked ticket) and so the simplest solution is just to stop
doing it.

See also discussion at:
https://bennettoxford.slack.com/archives/C069YDR4NCA/p1747925981081259

Closes #2466