Allow supplying dummy data directory instead of single file
This is only relevant to event level data support, but I noticed it as
part of the measures work as it will similarly need to support multuple
dummy data files.

The fix itself is very small, but it exposed an inadequacy in our
functional testing so I've added a test here as well.