Accept test data as a dict when only supplying single row
This contrasts with the previous behaviour which was to check the type
of the table and then _require_ either a single dict or a list of dicts
depending on whether the table was one-row-per-patient or not.

This caused confusion, particularly since doing it wrong led to a
traceback rather than a human-readable error message.