feat: Add ehrQL log file parser
This reads an ehrQL log file with Docker-style timestamps on stdin and
writes JSON lines to stdout. The JSON is intended to be suitable for
emiting as OpenTelemetry records. The tooling to do this was added to
the RAP Agent here:

 * https://github.com/opensafely-core/job-runner/pull/1088

The above PR explains more of the rationale for the particular approach
we've taken.

This is essentially spike work which was deemed useful enough to be
worth preserving, but not of high enough importance/urgency to fully
productionize right now. In particular, there aren't any tests so this
maay well break if we make significant changes to the log format.