Produce multiple column datasets in generative tests
We want to make sure we exercise the join splitting logic in the
Hypothesis tests because, through experience, this has the best chance
of catching weird corner-case issues that we've failed to think of.

More generally, it's probably a good idea to make our test datasets a
bit more realistic in this respect as there's a possibility that there
are bugs which can't be exposed in a single column dataset. Insofar as
we're interested in testing the per-variable logic then having more
variables per dataset should allow us to do so more efficiently. And
failure cases should (in theory) still shrink down to single column
examples where that's all that's needed to demonstrate the bug.

I've confirmed there's no appreciable difference in CI runtimes by doing
this.