Remind about user args syntax if there are unknown cli args
It's easy to forget that you need to put custom user args at the
end, and separate them with a --. And if you do, you get a
not-very-helpful "unrecognised arguments" error from argparse.

We can use parse_known_args to identify unknown args and show a
more helpful error message. Note that we also need to turn off
prefix matching, which should be fine, I think.