Modify generative serializer tests to use registry
I toyed with the idea of removing the serializer roundtrip checks from
the generative tests altogether. It was one of those things which was
added because it was very easy to do, rather than because we felt a
specific need to let Hypothesis lose on the serialization machinery.
However, on reflection, I think there is still value in letting
Hypothesis explore the edge cases here.

This does involve adding some extra production code, in the form of the
`deregister_object` fuction, which is only there for test purposes. But
it's simple enough code that this doesn't bother me too much.