Supporting adding metadata when defining tables
We do this via a Django-ish approach of defining an inner class. There
are other approaches we could take, but the big advantage of using a
class is that it makes it easy to include methods among the metadata
which I think we'll want to do for dummy data generation purposes.

In the first instance the only metadata property we support is supplying
a table name which differs from the class name.