Support introspecting measures to find all tables used
We need to be able to do this when enforcing permissions.

This is a slight improvement on the status quo I think in that we now at
least have a proper type to represent a collection of measures and some
consistency between `Dataset()` and `Measure()` in that both now have a
`_compile()` method.

Measures are still in a strange half-state in being a lot like query
model objects, but not actually being part of the query model. But
sorting that all out is beyond the scope of this PR.