Annotate all existing restricted tables
When referencing this in the user docs I've tried to use some suitable
generic text which doesn't suggest that you can just ask for permission
and expect to be granted it. We may well want to tweak this on a
per-table basis though, depending on the exact reason for the
restriction.