Check permissions are always included in table docs
I had previously tried to auto-generate the documentation here, but I
think we're better off having the flexibility to explain different
permissions in different ways. However I don't want us to be able to
forget to document them, or to typo the permission name, so we now
enforce this at documentation build time.