Raise a helpful error when using `abs()`
The `abs()` builtin is the Pythonic way of taking the absolute value,
but the inconsistency with other ehrQL methods is potentially confusing
for our users, most of whom are not experienced Python programmers.

We could support both options, but this risks confusion. So instead we
raise an error for anyone trying to use the builtin which points them in
the right direction.