feat: Update base docker image
The `latest` tag has been frozen on Ubuntu 20.04 as we've realised that
using `latest` is generally a bad idea but didn't want to break existing
downstream consumers.

Ubuntu 20.04 recently went out of mainstream support and so the Docker
image no longer build so we have to update.

Now we're no longer using `latest` we don't need the Hadolint ignore
pragma.