Temporarily ignore some Ruff upgrade rules
These change the syntax used for some generic types which we use to
power out autocompletion. The changes are probably fine, and the tests
seem to pass, but I'd rather not YOLO these right now because the type
signatures here are gnarly enough as it is.

I'll make a separate issue to look at these.