Remove unused T1OO config from tests
These were left in place by the below PR:
 * https://github.com/opensafely-core/ehrql/pull/2190

At the time there was a belief that the current approach to T1OO
filtering might be temporary and therefore that it made sense to leave
the test as they were. However, it's since been agreed to stick with the
current approach and so the presence of these tables in the tests is
both unnecessary and confusing.