Fix death date calculation in ehrQL tutorial
A patient is alive on the index date if their date of death is _after_ the
index date.

This didn't affect the tutorial output, because of a quirk of the dummy tables:
all the patients with a non-null death date also had a practice registration end
date which was before the index date. But that might not always be true of real
patient data.