Apply NDOO from permissions
National Data Opt Outs are applied to the dataset population in the
same way as T1OOs, by excluding any patients who appear in the
NationalDataOptOut table. Unlike T1OOs, we're using the new permissions
system to determine whether a particular ehrql query should include
them. We expect this permission to be provided in the EHRQL_PERMISSIONS
environment variable, which is supplied by the RAP controller and
applied by the RAP agent when running the job.