feat: Apply gp activation filtering
Apply the "V1B" filtering, as described in
https://docs.google.com/spreadsheets/d/1XkqNHMzUJKhvHvq59cpwm1_Dh2qyEszFRS2n8-BeE-E/edit?gid=0#gid=0
See also slack thread: https://bennettoxford.slack.com/archives/C069YDR4NCA/p1763383290662289?thread_ts=1763121414.807499&cid=C069YDR4NCA

Apply activation filtering overall and for GP events tables,
according to the following rules:

For practice_registrations table:
1) include ONLY activated registrations

For GP events tables, that is:
- appointments
- clinical_events,
- clinical_events_ranges
- medications
- repeat_medications
- vaccinations

1) No activated practice in registration history - exclude patient
   altogether
2) Patient is currently registered at an active practice or has
   deregistered from TPP altogether - include all event data
3) Patient is currently registered at an inactive practice: find
   the date of last deregistration from an active practice and
   include only events that occur on or before this date

For practice_registrations, we update the initial query to only
include activated registrations.

For event data, we apply transforms in modify_dataset; for each
applicable table, we build a new filtered version of the table,
that includes only those events that occur before the end date
in the activated table, and we replace any instances of the
table in the dataset query with the filtered version.