Apply gp activation filtering for clinical events
Apply activation filtering overall and for the clinical events table,
according to the following rules:

1) No activated practice in registration history - exclude patient
   altogether
2) Patient is currently registered at an active practice - include
   all clinical event data
3) Patient is currently registered at an inactive practice OR has
   deregistered: find the date of last deregistration from an
   active practice and filter out clinical events that occur
   after this date

This builds a new filtered clinical events table, that includes
only those events that occur before the end date in the activated
table.  It adds a new transform function, replace_source, which
replaces any instances of the clinical events table in the Dataset
nodes with the new activation-filtered clinical events table.