Add practice registration filtering based on GP activation
Practice registration data is also GP data so needs to be filtered.
However, it's slightly more complex because we know that sometimes
there are overlapping registrations, or registrations with the
same end date, or duplicate "current" registrations (ones that have
the same ceiling end date of 9999-12-31. So we can't just include
only practice registrations that end on or before after the last
activated date, because there could be some inactive ones that
end on the same last activated date.

So we filter the practice_registrations table to include only those
that end BEFORE the last activated date OR are activated.

This is a bit more complex currently than it needs to be because
we can't add the directions_acknowledged column to the actual
practice registrations table because it doesn't exist yet, so we
have an extra practice registrations table that includes it.