Mock the DirectionsAcknowledged column for performance testing
We want to test the performance of the gp activation filtering, but
we don't yet have the new column. We temporarily mock it based on the
Organisation_ID; this algorithm will return a determinate proportion
of activated practices (75%, which can be modified to test different
activation scenarios).

Temporarily create an organisation QueryTable to do this; in order to
test just the effect of the activation filtering, this is used in
all other tables that make use of the Organisation table
(practice_registrations, all_practice_registrations, activated). We
can test the same study code with and without activations by using
this branch with and without the relevant EHRQL_PERMISSIONS env
variable.